refactor: replace `test/index.ts` with `test/cpp.ts` file
