+++ src/test/cpp.ts
@@ -1,0 +1,2 @@
+export { compileCppTest } from "./cpp/compile.js";
+export { runCppTest } from "./cpp/run.js";
