--- src/test/cpp/index.ts
@@ -1,2 +1,0 @@
-export { compileCppTest } from "./compile.js";
-export { runCppTest } from "./run.js";
