--- src/test/index.ts
@@ -1,1 +1,0 @@
-export { compileCppTest, runCppTest } from "./cpp/index.js";
